Clear the old read_json function and replace it with the new read_json function
