+++ extension/json/CMakeLists.txt
@@ -15,7 +15,6 @@ set(JSON_EXTENSION_FILES
     json_common.cpp
     json_enums.cpp
     json_functions.cpp
-    json_multi_file_info.cpp
     json_scan.cpp
     json_serializer.cpp
     json_deserializer.cpp
@@ -44,7 +43,6 @@ set(JSON_EXTENSION_FILES
     json_functions/json_modify.cpp
     json_functions/json_strip_nulls.cpp
     json_functions/read_json.cpp
-    json_functions/read_json_objects.cpp
     json_functions/read_single_json_file.cpp)
 
 build_static_extension(json ${JSON_EXTENSION_FILES})
