+++ extension/json/include/json_functions.hpp
@@ -81,9 +81,8 @@ public:
 	static vector<TableFunctionSet> GetTableFunctions();
 	static unique_ptr<TableRef> ReadJSONReplacement(ClientContext &context, ReplacementScanInput &input,
 	                                                optional_ptr<ReplacementScanData> data);
-	static TableFunction GetReadJSONTableFunction(shared_ptr<JSONScanInfo> function_info);
+	static TableFunction GetJSONTableFunction(Identifier name, shared_ptr<JSONScanInfo> function_info);
 	static CopyFunction GetJSONCopyFunction();
-	static CopyFunction GetJSONNewCopyFunction();
 	static ScalarFunction GetJSONCopyToJSONFunction();
 	static ScalarFunction GetJSONCopyToGeoJSONFunction();
 	static CopyFunction GetGeoJSONCopyFunction();
@@ -159,8 +158,6 @@ private:
 	static TableFunctionSet GetReadNDJSONAutoFunction();
 	static TableFunction GetReadSingleJSONFileTableFunction(shared_ptr<JSONScanInfo> function_info);
 	static TableFunctionSet GetReadSingleJSONFileFunction();
-	static TableFunctionSet GetReadJSONNewFunction();
-	static TableFunction GetReadJSONNewTableFunction();
 
 	static TableFunctionSet GetJSONEachFunction();
 	static TableFunctionSet GetJSONTreeFunction();
