--- extension/json/include/json_multi_file_info.hpp
@@ -1,55 +1,0 @@
-//===----------------------------------------------------------------------===//
-//                         DuckDB
-//
-// json_multi_file_info.hpp
-//
-//
-//===----------------------------------------------------------------------===//
-
-#pragma once
-
-#include "duckdb/common/multi_file/multi_file_function.hpp"
-#include "json_reader_options.hpp"
-
-namespace duckdb {
-
-class JSONFileReaderOptions : public BaseFileReaderOptions {
-public:
-	JSONReaderOptions options;
-};
-
-struct JSONMultiFileInfo : MultiFileReaderInterface {
-	static unique_ptr<MultiFileReaderInterface> CreateInterface(ClientContext &context);
-
-	void InitializeFileOptions(MultiFileOptions &file_options) override;
-	unique_ptr<BaseFileReaderOptions> InitializeOptions(ClientContext &context,
-	                                                    optional_ptr<TableFunctionInfo> info) override;
-	bool ParseCopyOption(ClientContext &context, const Identifier &key, const vector<Value> &values,
-	                     BaseFileReaderOptions &options, vector<Identifier> &expected_names,
-	                     vector<LogicalType> &expected_types) override;
-	bool ParseOption(ClientContext &context, const Identifier &key, const Value &val, MultiFileOptions &file_options,
-	                 BaseFileReaderOptions &options) override;
-	void FinalizeCopyBind(ClientContext &context, BaseFileReaderOptions &options,
-	                      const vector<Identifier> &expected_names, const vector<LogicalType> &expected_types) override;
-	unique_ptr<TableFunctionData> InitializeBindData(MultiFileBindData &multi_file_data,
-	                                                 unique_ptr<BaseFileReaderOptions> options) override;
-	void BindReader(ClientContext &context, vector<LogicalType> &return_types, vector<Identifier> &names,
-	                MultiFileBindData &bind_data) override;
-	optional_idx MaxThreads(const MultiFileBindData &bind_data, const MultiFileGlobalState &global_state,
-	                        FileExpandResult expand_result) override;
-	unique_ptr<GlobalTableFunctionState> InitializeGlobalState(ClientContext &context, MultiFileBindData &bind_data,
-	                                                           MultiFileGlobalState &global_state) override;
-	unique_ptr<LocalTableFunctionState> InitializeLocalState(ClientContext &context,
-	                                                         GlobalTableFunctionState &global_state) override;
-	shared_ptr<BaseFileReader> CreateReader(ClientContext &context, GlobalTableFunctionState &gstate,
-	                                        BaseUnionData &union_data, const MultiFileBindData &bind_data_p) override;
-	shared_ptr<BaseFileReader> CreateReader(ClientContext &context, GlobalTableFunctionState &gstate,
-	                                        const OpenFileInfo &file, idx_t file_idx,
-	                                        const MultiFileBindData &bind_data) override;
-	void FinishReading(ClientContext &context, GlobalTableFunctionState &global_state,
-	                   LocalTableFunctionState &local_state) override;
-	unique_ptr<NodeStatistics> GetCardinality(const MultiFileBindData &bind_data, idx_t file_count) override;
-	FileGlobInput GetGlobInput() override;
-};
-
-} // namespace duckdb
