+++ extension/json/json_functions/copy_json.cpp
@@ -16,7 +16,6 @@
 #include "json_functions.hpp"
 #include "json_scan.hpp"
 #include "json_transform.hpp"
-#include "json_multi_file_info.hpp"
 #include "duckdb/common/multi_file/table_function_multi_file.hpp"
 #include "duckdb/parser/expression/cast_expression.hpp"
 
@@ -335,29 +334,22 @@ static BoundStatement CopyToGeoJSONPlan(Binder &binder, CopyStatement &stmt) {
 	return CopyToJSONPlanInternal(binder, stmt, JSONCopyToFormat::GEOJSON);
 }
 
+//! COPY ... FROM reads through the multi-file wrapper around read_single_json_file. COPY takes its columns from
+//! the target table, so the reader binds every file against those columns instead of auto-detecting a schema
+static void SetJSONCopyFromFunction(CopyFunction &function, Identifier name) {
+	function.copy_from_bind = TableFunctionMultiFileWrapper::MultiFileBindCopy;
+	function.copy_from_function = JSONFunctions::GetJSONTableFunction(
+	    std::move(name), make_shared_ptr<JSONScanInfo>(JSONScanType::READ_JSON, JSONFormat::AUTO_DETECT,
+	                                                   JSONRecordType::RECORDS, false));
+}
+
 CopyFunction JSONFunctions::GetGeoJSONCopyFunction() {
 	CopyFunction function("geojson");
 	function.extension = "geojson";
 
 	function.plan = CopyToGeoJSONPlan;
 
-	function.copy_from_bind = MultiFileFunction<JSONMultiFileInfo>::MultiFileBindCopy;
-	function.copy_from_function = JSONFunctions::GetReadJSONTableFunction(make_shared_ptr<JSONScanInfo>(
-	    JSONScanType::READ_JSON, JSONFormat::AUTO_DETECT, JSONRecordType::RECORDS, false));
-
-	return function;
-}
-
-//! The COPY function of read_json_new - reading is done by the multi-file wrapper around read_single_json_file,
-//! writing is shared with the regular JSON copy function
-CopyFunction JSONFunctions::GetJSONNewCopyFunction() {
-	CopyFunction function("json_new");
-	function.extension = "json";
-
-	function.plan = CopyToJSONPlan;
-
-	function.copy_from_bind = TableFunctionMultiFileWrapper::MultiFileBindCopy;
-	function.copy_from_function = JSONFunctions::GetReadJSONNewTableFunction();
+	SetJSONCopyFromFunction(function, "read_json");
 
 	return function;
 }
@@ -367,10 +359,7 @@ CopyFunction JSONFunctions::GetJSONCopyFunction() {
 	function.extension = "json";
 
 	function.plan = CopyToJSONPlan;
-
-	function.copy_from_bind = MultiFileFunction<JSONMultiFileInfo>::MultiFileBindCopy;
-	function.copy_from_function = JSONFunctions::GetReadJSONTableFunction(make_shared_ptr<JSONScanInfo>(
-	    JSONScanType::READ_JSON, JSONFormat::AUTO_DETECT, JSONRecordType::RECORDS, false));
+	SetJSONCopyFromFunction(function, "read_json");
 
 	return function;
 }
