--- extension/json/json_multi_file_info.cpp
@@ -1,695 +1,0 @@
-#include "json_multi_file_info.hpp"
-#include "duckdb/main/client_context.hpp"
-#include "json_scan.hpp"
-#include "duckdb/common/types/value.hpp"
-#include "duckdb/parallel/async_result.hpp"
-
-namespace duckdb {
-
-unique_ptr<MultiFileReaderInterface> JSONMultiFileInfo::CreateInterface(ClientContext &context) {
-	return make_uniq<JSONMultiFileInfo>();
-}
-
-void JSONMultiFileInfo::InitializeFileOptions(MultiFileOptions &file_options) {
-	// read_json determines the schema by combining the schemas of up to 32 files
-	file_options.maximum_sample_files = 32;
-}
-
-unique_ptr<BaseFileReaderOptions> JSONMultiFileInfo::InitializeOptions(ClientContext &context,
-                                                                       optional_ptr<TableFunctionInfo> info) {
-	auto reader_options = make_uniq<JSONFileReaderOptions>();
-	auto &options = reader_options->options;
-	if (info) {
-		auto &scan_info = info->Cast<JSONScanInfo>();
-		options.type = scan_info.type;
-		options.format = scan_info.format;
-		options.record_type = scan_info.record_type;
-		options.auto_detect = scan_info.auto_detect;
-		if (scan_info.type == JSONScanType::READ_JSON_OBJECTS) {
-			// read_json_objects always emits a single JSON column called "json"
-			options.sql_type_list.push_back(LogicalType::JSON());
-			options.name_list.emplace_back("json");
-		}
-	} else {
-		// COPY
-		options.type = JSONScanType::READ_JSON;
-		options.record_type = JSONRecordType::RECORDS;
-		options.format = JSONFormat::AUTO_DETECT;
-		options.auto_detect = false;
-	}
-	return std::move(reader_options);
-}
-
-bool JSONMultiFileInfo::ParseOption(ClientContext &context, const Identifier &key, const Value &value,
-                                    MultiFileOptions &, BaseFileReaderOptions &options_p) {
-	return JSONScan::ParseOption(context, key, value, options_p.Cast<JSONFileReaderOptions>().options);
-}
-
-bool JSONScan::ParseOption(ClientContext &context, const Identifier &key, const Value &value,
-                           JSONReaderOptions &options) {
-	if (value.IsNull()) {
-		throw BinderException("Cannot use NULL as argument to key %s", key);
-	}
-	if (key == "ignore_errors") {
-		options.ignore_errors = BooleanValue::Get(value);
-		return true;
-	}
-	if (key == "maximum_object_size") {
-		options.maximum_object_size = MaxValue<idx_t>(UIntegerValue::Get(value), options.maximum_object_size);
-		return true;
-	}
-	if (key == "format") {
-		auto arg = StringUtil::Lower(StringValue::Get(value));
-		static const auto FORMAT_OPTIONS =
-		    case_insensitive_map_t<JSONFormat> {{"auto", JSONFormat::AUTO_DETECT},
-		                                        {"unstructured", JSONFormat::UNSTRUCTURED},
-		                                        {"newline_delimited", JSONFormat::NEWLINE_DELIMITED},
-		                                        {"nd", JSONFormat::NEWLINE_DELIMITED},
-		                                        {"array", JSONFormat::ARRAY}};
-		auto lookup = FORMAT_OPTIONS.find(arg);
-		if (lookup == FORMAT_OPTIONS.end()) {
-			vector<string> valid_options;
-			for (auto &pair : FORMAT_OPTIONS) {
-				valid_options.push_back(StringUtil::Format("'%s'", pair.first));
-			}
-			throw BinderException("format must be one of [%s], not '%s'", StringUtil::Join(valid_options, ", "), arg);
-		}
-		options.format = lookup->second;
-		return true;
-	}
-	if (key == "compression") {
-		options.compression = FileCompressionType(StringValue::Get(value));
-		return true;
-	}
-	if (key == "columns") {
-		auto &child_type = value.type();
-		if (child_type.id() != LogicalTypeId::STRUCT) {
-			throw BinderException("read_json \"columns\" parameter requires a struct as input.");
-		}
-		auto &struct_children = StructValue::GetChildren(value);
-		D_ASSERT(StructType::GetChildCount(child_type) == struct_children.size());
-		for (idx_t i = 0; i < struct_children.size(); i++) {
-			auto &name = StructType::GetChildName(child_type, i);
-			auto &val = struct_children[i];
-			if (val.IsNull()) {
-				throw BinderException("read_json \"columns\" parameter type specification cannot be NULL.");
-			}
-			options.name_list.emplace_back(name);
-			if (val.type().id() != LogicalTypeId::VARCHAR) {
-				throw BinderException("read_json \"columns\" parameter type specification must be VARCHAR.");
-			}
-			options.sql_type_list.emplace_back(TransformStringToLogicalType(StringValue::Get(val), context));
-		}
-		D_ASSERT(options.name_list.size() == options.sql_type_list.size());
-		if (options.name_list.empty()) {
-			throw BinderException("read_json \"columns\" parameter needs at least one column.");
-		}
-		return true;
-	}
-	if (key == "auto_detect") {
-		options.auto_detect = BooleanValue::Get(value);
-		return true;
-	}
-	if (key == "geojson") {
-		options.geojson = BooleanValue::Get(value);
-		return true;
-	}
-	if (key == "sample_size") {
-		auto arg = BigIntValue::Get(value);
-		if (arg == -1) {
-			options.sample_size = NumericLimits<idx_t>::Maximum();
-		} else if (arg > 0) {
-			options.sample_size = arg;
-		} else {
-			throw BinderException("read_json \"sample_size\" parameter must be positive, or -1 to sample all input "
-			                      "files entirely, up to \"maximum_sample_files\" files.");
-		}
-		return true;
-	}
-	if (key == "maximum_depth") {
-		auto arg = BigIntValue::Get(value);
-		if (arg == -1) {
-			options.max_depth = NumericLimits<idx_t>::Maximum();
-		} else {
-			options.max_depth = arg;
-		}
-		return true;
-	}
-	if (key == "field_appearance_threshold") {
-		auto arg = DoubleValue::Get(value);
-		if (arg < 0 || arg > 1) {
-			throw BinderException("read_json_auto \"field_appearance_threshold\" parameter must be between 0 and 1");
-		}
-		options.field_appearance_threshold = arg;
-		return true;
-	}
-	if (key == "map_inference_threshold") {
-		auto arg = BigIntValue::Get(value);
-		if (arg == -1) {
-			options.map_inference_threshold = NumericLimits<idx_t>::Maximum();
-		} else if (arg >= 0) {
-			options.map_inference_threshold = arg;
-		} else {
-			throw BinderException("read_json_auto \"map_inference_threshold\" parameter must be 0 or positive, "
-			                      "or -1 to disable map inference for consistent objects.");
-		}
-		return true;
-	}
-	if (key == "dateformat" || key == "date_format") {
-		auto format_string = StringValue::Get(value);
-		if (StringUtil::Lower(format_string) == "iso") {
-			format_string = "%Y-%m-%d";
-		}
-		options.date_format = format_string;
-
-		StrpTimeFormat format;
-		auto error = StrTimeFormat::ParseFormatSpecifier(format_string, format);
-		if (!error.empty()) {
-			throw BinderException("read_json could not parse \"dateformat\": '%s'.", error.c_str());
-		}
-		return true;
-	}
-	if (key == "timestampformat" || key == "timestamp_format") {
-		auto format_string = StringValue::Get(value);
-		if (StringUtil::Lower(format_string) == "iso") {
-			format_string = "%Y-%m-%dT%H:%M:%S.%fZ";
-		}
-		options.timestamp_format = format_string;
-
-		StrpTimeFormat format;
-		auto error = StrTimeFormat::ParseFormatSpecifier(format_string, format);
-		if (!error.empty()) {
-			throw BinderException("read_json could not parse \"timestampformat\": '%s'.", error.c_str());
-		}
-		return true;
-	}
-	if (key == "records") {
-		auto arg = StringValue::Get(value);
-		if (arg == "auto") {
-			options.record_type = JSONRecordType::AUTO_DETECT;
-		} else if (arg == "true") {
-			options.record_type = JSONRecordType::RECORDS;
-		} else if (arg == "false") {
-			options.record_type = JSONRecordType::VALUES;
-		} else {
-			throw BinderException("read_json requires \"records\" to be one of ['auto', 'true', 'false'].");
-		}
-		return true;
-	}
-	if (key == "maximum_sample_files") {
-		auto arg = BigIntValue::Get(value);
-		if (arg == -1) {
-			options.maximum_sample_files = NumericLimits<idx_t>::Maximum();
-		} else if (arg > 0) {
-			options.maximum_sample_files = arg;
-		} else {
-			throw BinderException("read_json \"maximum_sample_files\" parameter must be positive, or -1 to remove "
-			                      "the limit on the number of files used to sample \"sample_size\" rows.");
-		}
-		return true;
-	}
-	if (key == "convert_strings_to_integers") {
-		options.convert_strings_to_integers = BooleanValue::Get(value);
-		return true;
-	}
-	return false;
-}
-
-static void JSONCheckSingleParameter(const Identifier &key, const vector<Value> &values) {
-	if (values.size() == 1) {
-		return;
-	}
-	throw BinderException("COPY (FORMAT JSON) parameter %s expects a single argument.", key);
-}
-
-bool JSONMultiFileInfo::ParseCopyOption(ClientContext &context, const Identifier &key, const vector<Value> &values,
-                                        BaseFileReaderOptions &options_p, vector<Identifier> &expected_names,
-                                        vector<LogicalType> &expected_types) {
-	auto &reader_options = options_p.Cast<JSONFileReaderOptions>();
-	auto &options = reader_options.options;
-	if (key == "dateformat" || key == "date_format") {
-		JSONCheckSingleParameter(key, values);
-		options.date_format = StringValue::Get(values.back());
-		return true;
-	}
-	if (key == "timestampformat" || key == "timestamp_format") {
-		JSONCheckSingleParameter(key, values);
-		options.timestamp_format = StringValue::Get(values.back());
-		return true;
-	}
-	if (key == "auto_detect") {
-		if (values.empty()) {
-			options.auto_detect = true;
-		} else {
-			JSONCheckSingleParameter(key, values);
-			options.auto_detect = BooleanValue::Get(values.back().DefaultCastAs(LogicalTypeId::BOOLEAN));
-			if (options.format == JSONFormat::AUTO_DETECT) {
-				options.format = JSONFormat::NEWLINE_DELIMITED;
-			}
-		}
-		return true;
-	}
-	if (key == "geojson") {
-		if (values.empty()) {
-			options.geojson = true;
-		} else {
-			JSONCheckSingleParameter(key, values);
-			options.geojson = BooleanValue::Get(values.back().DefaultCastAs(LogicalTypeId::BOOLEAN));
-		}
-		return true;
-	}
-	if (key == "compression") {
-		JSONCheckSingleParameter(key, values);
-		options.compression = FileCompressionType(StringValue::Get(values.back()));
-		return true;
-	}
-	if (key == "array") {
-		if (values.empty()) {
-			options.format = JSONFormat::ARRAY;
-		} else {
-			JSONCheckSingleParameter(key, values);
-			if (BooleanValue::Get(values.back().DefaultCastAs(LogicalTypeId::BOOLEAN))) {
-				options.format = JSONFormat::ARRAY;
-			} else {
-				// Default to newline-delimited otherwise
-				options.format = JSONFormat::NEWLINE_DELIMITED;
-			}
-		}
-		return true;
-	}
-	return false;
-}
-
-unique_ptr<TableFunctionData> JSONMultiFileInfo::InitializeBindData(MultiFileBindData &multi_file_data,
-                                                                    unique_ptr<BaseFileReaderOptions> options) {
-	auto &reader_options = options->Cast<JSONFileReaderOptions>();
-	auto json_data = make_uniq<JSONScanData>();
-	json_data->options = std::move(reader_options.options);
-	return std::move(json_data);
-}
-
-void JSONScan::BindSchema(ClientContext &context, JSONScanData &json_data, MultiFileList &files,
-                          vector<shared_ptr<BaseUnionData>> &union_readers, bool union_by_name,
-                          vector<LogicalType> &return_types, vector<Identifier> &names) {
-	auto &options = json_data.options;
-	names = options.name_list;
-	return_types = options.sql_type_list;
-	if (options.record_type == JSONRecordType::AUTO_DETECT && return_types.size() > 1) {
-		// More than one specified column implies records
-		options.record_type = JSONRecordType::RECORDS;
-	}
-
-	// Specifying column names overrides auto-detect
-	if (!return_types.empty()) {
-		options.auto_detect = false;
-	}
-
-	if (!options.auto_detect) {
-		// Need to specify columns if RECORDS and not auto-detecting
-		if (return_types.empty()) {
-			throw BinderException("When auto_detect=false, read_json requires columns to be specified through the "
-			                      "\"columns\" parameter.");
-		}
-		// If we are reading VALUES, we can only have one column
-		if (options.record_type == JSONRecordType::VALUES && return_types.size() != 1) {
-			throw BinderException("read_json requires a single column to be specified through the \"columns\" "
-			                      "parameter when \"records\" is set to 'false'.");
-		}
-	}
-
-	// Reading GeoJSON is opt-in, but a .geojson / .geojsonl file name is opt-in enough
-	if (!options.geojson.has_value()) {
-		const auto first_file = files.GetFirstFile().path;
-		options.geojson =
-		    StringUtil::CIEndsWith(first_file, ".geojson") || StringUtil::CIEndsWith(first_file, ".geojsonl");
-	}
-	// COPY ... FROM hardcodes RECORDS because it takes its columns from the target table. For GeoJSON we still
-	// need to know whether those columns live in a Feature's "properties", so let the structure decide.
-	if (options.geojson.value() && !options.auto_detect && !options.name_list.empty() &&
-	    options.record_type == JSONRecordType::RECORDS) {
-		options.record_type = JSONRecordType::AUTO_DETECT;
-	}
-
-	json_data.InitializeFormats();
-
-	if (options.auto_detect || options.record_type == JSONRecordType::AUTO_DETECT) {
-		JSONScan::AutoDetect(context, json_data, files.GetAllFiles(), union_readers, union_by_name, return_types,
-		                     names);
-		D_ASSERT(return_types.size() == names.size());
-	}
-	json_data.key_names = IdentifiersToStrings(names);
-}
-
-void JSONScan::FinalizeBind(JSONScanData &json_data, vector<Identifier> &names) {
-	auto &options = json_data.options;
-	auto &transform_options = json_data.transform_options;
-	transform_options.strict_cast = !options.ignore_errors;
-	transform_options.error_duplicate_key = !options.ignore_errors;
-	transform_options.error_missing_key = false;
-	transform_options.error_unknown_key = options.auto_detect && !options.ignore_errors;
-	transform_options.date_format_map = json_data.date_format_map.get();
-	transform_options.delay_error = true;
-
-	if (!options.auto_detect) {
-		return;
-	}
-	// Note that we can't change json_data.key_names, because the JSON reader gets columns by exact name, not position
-	DeduplicateColumnNames(names);
-}
-
-void JSONScan::DeduplicateColumnNames(vector<Identifier> &names) {
-	identifier_map_t<idx_t> name_collision_count;
-	for (auto &col_name : names) {
-		// Taken from CSV header_detection.cpp
-		while (name_collision_count.find(col_name) != name_collision_count.end()) {
-			name_collision_count[col_name] += 1;
-			col_name = Identifier(col_name + "_" + to_string(name_collision_count[col_name]));
-		}
-		name_collision_count[col_name] = 0;
-	}
-}
-
-void JSONMultiFileInfo::BindReader(ClientContext &context, vector<LogicalType> &return_types, vector<Identifier> &names,
-                                   MultiFileBindData &bind_data) {
-	auto &json_data = bind_data.bind_data->Cast<JSONScanData>();
-	// "maximum_sample_files" is a multi file option - the JSON auto-detection samples that many files
-	json_data.options.maximum_sample_files = bind_data.file_options.maximum_sample_files;
-
-	JSONScan::BindSchema(context, json_data, *bind_data.file_list, bind_data.union_readers,
-	                     bind_data.file_options.union_by_name, return_types, names);
-
-	bind_data.multi_file_reader->BindOptions(bind_data.file_options, *bind_data.file_list, return_types, names,
-	                                         bind_data.reader_bind);
-
-	JSONScan::FinalizeBind(json_data, names);
-
-	bool reuse_readers = true;
-	for (auto &union_reader : bind_data.union_readers) {
-		if (!union_reader || !union_reader->reader) {
-			// not all readers have been initialized - don't re-use any
-			reuse_readers = false;
-			break;
-		}
-		auto &json_reader = union_reader->reader->Cast<JSONReader>();
-		if (!json_reader.IsOpen()) {
-			// no open file-handle - close
-			reuse_readers = false;
-		}
-	}
-	if (!reuse_readers) {
-		bind_data.union_readers.clear();
-	} else {
-		// re-use readers
-		for (auto &union_reader : bind_data.union_readers) {
-			auto &json_reader = union_reader->reader->Cast<JSONReader>();
-			union_reader->names = IdentifiersToStrings(names);
-			union_reader->types = return_types;
-			union_reader->reader->columns = MultiFileColumnDefinition::ColumnsFromNamesAndTypes(names, return_types);
-			json_reader.Reset();
-		}
-	}
-}
-
-void JSONMultiFileInfo::FinalizeCopyBind(ClientContext &context, BaseFileReaderOptions &options_p,
-                                         const vector<Identifier> &expected_names,
-                                         const vector<LogicalType> &expected_types) {
-	auto &reader_options = options_p.Cast<JSONFileReaderOptions>();
-	auto &options = reader_options.options;
-	options.name_list = expected_names;
-	options.sql_type_list = expected_types;
-	if (options.auto_detect && options.format != JSONFormat::ARRAY) {
-		options.format = JSONFormat::AUTO_DETECT;
-	}
-}
-
-unique_ptr<GlobalTableFunctionState> JSONMultiFileInfo::InitializeGlobalState(ClientContext &context,
-                                                                              MultiFileBindData &bind_data,
-                                                                              MultiFileGlobalState &global_state) {
-	auto &json_data = bind_data.bind_data->Cast<JSONScanData>();
-	auto json_state =
-	    make_uniq<JSONGlobalTableFunctionState>(context, json_data, bind_data.file_list->GetTotalFileCount());
-
-	auto &gstate = json_state->state;
-	// Perform projection pushdown
-	for (idx_t col_idx = 0; col_idx < global_state.column_indexes.size(); col_idx++) {
-		auto &column_index = global_state.column_indexes[col_idx];
-		const auto &col_id = column_index.GetPrimaryIndex();
-
-		// Skip any multi-file reader / row id stuff
-		if (bind_data.reader_bind.filename_idx.IsValid() && col_id == bind_data.reader_bind.filename_idx.GetIndex()) {
-			continue;
-		}
-		if (IsVirtualColumn(col_id)) {
-			continue;
-		}
-		bool skip = false;
-		for (const auto &hive_partitioning_index : bind_data.reader_bind.hive_partitioning_indexes) {
-			if (col_id == hive_partitioning_index.index) {
-				skip = true;
-				break;
-			}
-		}
-		if (skip) {
-			continue;
-		}
-
-		gstate.names.push_back(json_data.key_names[col_id]);
-		gstate.column_ids.push_back(col_idx);
-		gstate.column_indices.push_back(column_index);
-		if (!json_data.feature_columns.empty()) {
-			gstate.feature_columns.push_back(json_data.feature_columns[col_id]);
-		}
-	}
-	if (gstate.names.size() < json_data.key_names.size() || bind_data.file_options.union_by_name) {
-		// If we are auto-detecting, but don't need all columns present in the file,
-		// then we don't need to throw an error if we encounter an unseen column
-		gstate.transform_options.error_unknown_key = false;
-	}
-	return std::move(json_state);
-}
-
-unique_ptr<LocalTableFunctionState> JSONMultiFileInfo::InitializeLocalState(ClientContext &context,
-                                                                            GlobalTableFunctionState &global_state) {
-	auto &gstate = global_state.Cast<JSONGlobalTableFunctionState>();
-	auto result = make_uniq<JSONLocalTableFunctionState>(context, gstate.state);
-
-	// Copy the transform options / date format map because we need to do thread-local stuff
-	result->state.transform_options = gstate.state.transform_options;
-
-	return std::move(result);
-}
-
-double JSONReader::GetProgressInFile(ClientContext &context) {
-	return GetProgress();
-}
-
-shared_ptr<BaseFileReader> JSONMultiFileInfo::CreateReader(ClientContext &context, GlobalTableFunctionState &gstate_p,
-                                                           BaseUnionData &union_data,
-                                                           const MultiFileBindData &bind_data_p) {
-	auto &json_data = bind_data_p.bind_data->Cast<JSONScanData>();
-	auto reader = make_shared_ptr<JSONReader>(context, json_data.options, union_data.GetFileName());
-	reader->columns =
-	    MultiFileColumnDefinition::ColumnsFromNamesAndTypes(StringsToIdentifiers(union_data.names), union_data.types);
-	return std::move(reader);
-}
-
-shared_ptr<BaseFileReader> JSONMultiFileInfo::CreateReader(ClientContext &context, GlobalTableFunctionState &gstate_p,
-                                                           const OpenFileInfo &file, idx_t file_idx,
-                                                           const MultiFileBindData &bind_data) {
-	auto &json_data = bind_data.bind_data->Cast<JSONScanData>();
-	auto reader = make_shared_ptr<JSONReader>(context, json_data.options, file.path);
-	reader->columns = MultiFileColumnDefinition::ColumnsFromNamesAndTypes(bind_data.names, bind_data.types);
-	return std::move(reader);
-}
-
-void JSONReader::PrepareReader(ClientContext &context, GlobalTableFunctionState &gstate_p) {
-	auto &gstate = gstate_p.Cast<JSONGlobalTableFunctionState>().state;
-	if (gstate.enable_parallel_scans) {
-		// if we are doing parallel scans we need to open the file here
-		Initialize(gstate.allocator, gstate.buffer_capacity);
-	}
-}
-
-bool JSONReader::TryInitializeScan(ClientContext &context, GlobalTableFunctionState &gstate_p,
-                                   LocalTableFunctionState &lstate_p) {
-	auto &gstate = gstate_p.Cast<JSONGlobalTableFunctionState>().state;
-	auto &lstate = lstate_p.Cast<JSONLocalTableFunctionState>().state;
-
-	lstate.GetScanState().ResetForNextBuffer();
-	return lstate.TryInitializeScan(gstate, *this);
-}
-
-//! GeoJSON Features are read as two groups of columns: those taken from the Feature itself, and those taken from
-//! its "properties" object
-static bool TransformGeoJSONFeatures(yyjson_val *values[], yyjson_alc *alc, const idx_t count,
-                                     JSONScanGlobalState &gstate, const vector<Vector *> &result_vectors,
-                                     JSONTransformOptions &transform_options) {
-	D_ASSERT(gstate.feature_columns.size() == gstate.names.size());
-
-	// Group the columns by where they read from, using the JSON key rather than the (deduplicated) column name.
-	// A Feature member and a property of the same name are separate columns, so this cannot be done by name.
-	vector<string> feature_names, property_names;
-	vector<Vector *> feature_vectors, property_vectors;
-	vector<ColumnIndex> feature_indices, property_indices;
-	for (idx_t i = 0; i < gstate.feature_columns.size(); i++) {
-		const auto &feature_column = gstate.feature_columns[i];
-		const auto from_properties = feature_column.from_properties;
-		(from_properties ? property_names : feature_names).push_back(feature_column.key);
-		(from_properties ? property_vectors : feature_vectors).push_back(result_vectors[i]);
-		if (!gstate.column_indices.empty()) {
-			(from_properties ? property_indices : feature_indices).push_back(gstate.column_indices[i]);
-		}
-	}
-
-	// "type", "properties" and "bbox" are consumed by the unnesting, so unknown keys are expected in both groups
-	bool success = true;
-	if (!feature_names.empty()) {
-		success = JSONTransform::TransformObject(values, alc, count, feature_names, feature_vectors, transform_options,
-		                                         feature_indices, false);
-	}
-	if (success && !property_names.empty()) {
-		vector<yyjson_val *> property_values(count);
-		for (idx_t i = 0; i < count; i++) {
-			auto &val = values[i];
-			property_values[i] = val && unsafe_yyjson_is_obj(val) ? yyjson_obj_get(val, "properties") : nullptr;
-		}
-		success = JSONTransform::TransformObject(property_values.data(), alc, count, property_names, property_vectors,
-		                                         transform_options, property_indices, false);
-	}
-	return success;
-}
-
-void ReadJSONFunction(ClientContext &context, JSONReader &json_reader, JSONScanGlobalState &gstate,
-                      JSONScanLocalState &lstate, DataChunk &output) {
-	auto &scan_state = lstate.GetScanState();
-	D_ASSERT(RefersToSameObject(json_reader, *scan_state.current_reader));
-
-	const auto count = lstate.Read();
-	yyjson_val **values = scan_state.values;
-
-	if (!gstate.names.empty()) {
-		vector<Vector *> result_vectors;
-		result_vectors.reserve(gstate.names.size());
-		for (idx_t i = 0; i < gstate.names.size(); i++) {
-			result_vectors.emplace_back(&output.data[i]);
-		}
-
-		D_ASSERT(gstate.json_data.options.record_type != JSONRecordType::AUTO_DETECT);
-		bool success;
-		if (gstate.json_data.options.record_type == JSONRecordType::RECORDS) {
-			success = JSONTransform::TransformObject(values, scan_state.allocator.GetYYAlc(), count, gstate.names,
-			                                         result_vectors, lstate.transform_options, gstate.column_indices,
-			                                         lstate.transform_options.error_unknown_key);
-		} else if (gstate.json_data.options.record_type == JSONRecordType::FEATURES) {
-			success = TransformGeoJSONFeatures(values, scan_state.allocator.GetYYAlc(), count, gstate, result_vectors,
-			                                   lstate.transform_options);
-		} else {
-			D_ASSERT(gstate.json_data.options.record_type == JSONRecordType::VALUES);
-			success = JSONTransform::Transform(values, scan_state.allocator.GetYYAlc(), *result_vectors[0], count,
-			                                   lstate.transform_options, gstate.column_indices[0]);
-		}
-
-		if (!success) {
-			string hint =
-			    gstate.json_data.options.auto_detect
-			        ? "\nTry increasing 'sample_size', reducing 'maximum_depth', specifying 'columns', 'format' or "
-			          "'records' manually, setting 'ignore_errors' to true, or setting 'union_by_name' to true when "
-			          "reading multiple files with a different structure."
-			        : "\nTry setting 'auto_detect' to true, specifying 'format' or 'records' manually, or setting "
-			          "'ignore_errors' to true.";
-			lstate.AddTransformError(lstate.transform_options.object_index,
-			                         lstate.transform_options.error_message + hint);
-			return;
-		}
-	}
-	output.SetChildCardinality(count);
-}
-
-void ReadJSONObjectsFunction(ClientContext &context, JSONReader &json_reader, JSONScanGlobalState &gstate,
-                             JSONScanLocalState &lstate, DataChunk &output) {
-	// Fetch next lines
-	auto &scan_state = lstate.GetScanState();
-	D_ASSERT(RefersToSameObject(json_reader, *scan_state.current_reader));
-
-	const auto count = lstate.Read();
-	const auto units = scan_state.units;
-	const auto objects = scan_state.values;
-
-	if (!gstate.names.empty()) {
-		// Create the strings without copying them
-		auto strings = FlatVector::Writer<string_t>(output.data[0], count);
-		for (idx_t i = 0; i < count; i++) {
-			if (objects[i]) {
-				strings.WriteStringRef(string_t(units[i].pointer, units[i].size));
-			} else {
-				strings.WriteNull();
-			}
-		}
-	}
-
-	output.SetChildCardinality(count);
-}
-
-AsyncResult JSONReader::Scan(ClientContext &context, GlobalTableFunctionState &global_state,
-                             LocalTableFunctionState &local_state, DataChunk &output) {
-#ifdef DUCKDB_DEBUG_ASYNC_SINK_SOURCE
-	{
-		AsyncResult test_result;
-		if (AsyncResult::TryGenerateTestResult(test_result)) {
-			return test_result;
-		}
-	}
-#endif
-
-	auto &gstate = global_state.Cast<JSONGlobalTableFunctionState>().state;
-	auto &lstate = local_state.Cast<JSONLocalTableFunctionState>().state;
-	auto &json_data = gstate.json_data;
-	switch (json_data.options.type) {
-	case JSONScanType::READ_JSON:
-		ReadJSONFunction(context, *this, gstate, lstate, output);
-		break;
-	case JSONScanType::READ_JSON_OBJECTS:
-		ReadJSONObjectsFunction(context, *this, gstate, lstate, output);
-		break;
-	default:
-		throw InternalException("Unsupported scan type for JSONMultiFileInfo::Scan");
-	}
-	return AsyncResult::FromChunk(output);
-}
-
-void JSONReader::FinishFile(ClientContext &context, GlobalTableFunctionState &global_state) {
-	auto &gstate = global_state.Cast<JSONGlobalTableFunctionState>().state;
-	gstate.file_is_assigned = false;
-}
-
-void JSONMultiFileInfo::FinishReading(ClientContext &context, GlobalTableFunctionState &global_state,
-                                      LocalTableFunctionState &local_state) {
-	auto &lstate = local_state.Cast<JSONLocalTableFunctionState>().state;
-	lstate.GetScanState().ResetForNextBuffer();
-}
-
-unique_ptr<NodeStatistics> JSONMultiFileInfo::GetCardinality(const MultiFileBindData &bind_data, idx_t file_count) {
-	auto &json_data = bind_data.bind_data->Cast<JSONScanData>();
-	idx_t per_file_cardinality = 42;
-	// get the average per-file cardinality from the bind data (if it is set)
-	if (json_data.estimated_cardinality_per_file.IsValid()) {
-		per_file_cardinality = json_data.estimated_cardinality_per_file.GetIndex();
-	}
-	return make_uniq<NodeStatistics>(per_file_cardinality * file_count);
-}
-
-optional_idx JSONMultiFileInfo::MaxThreads(const MultiFileBindData &bind_data, const MultiFileGlobalState &global_state,
-                                           FileExpandResult expand_result) {
-	if (expand_result == FileExpandResult::MULTIPLE_FILES) {
-		return optional_idx();
-	}
-	// get the max threads from the bind data (if it is set)
-	auto &json_data = bind_data.bind_data->Cast<JSONScanData>();
-	return json_data.max_threads;
-}
-
-FileGlobInput JSONMultiFileInfo::GetGlobInput() {
-	return FileGlobInput(FileGlobOptions::FALLBACK_GLOB, "json");
-}
-
-} // namespace duckdb
