+++ extension/json/json_reader.cpp
@@ -179,8 +179,8 @@ idx_t JSONFileHandle::ReadFromCache(char *&pointer, idx_t &size, atomic<idx_t> &
 }
 
 JSONReader::JSONReader(ClientContext &context, JSONReaderOptions options_p, OpenFileInfo file_p)
-    : BaseFileReader(std::move(file_p)), context(context), options(std::move(options_p)), initialized(0),
-      next_buffer_index(0), thrown(false) {
+    : file(std::move(file_p)), context(context), options(std::move(options_p)), initialized(0), next_buffer_index(0),
+      thrown(false) {
 }
 
 void JSONReader::OpenJSONFile() {
