+++ extension/json/json_scan.cpp
@@ -8,8 +8,9 @@
 #include "duckdb/main/extension_helper.hpp"
 #include "duckdb/parallel/task_scheduler.hpp"
 #include "duckdb/storage/buffer_manager.hpp"
-#include "json_multi_file_info.hpp"
 #include "json_structure.hpp"
+#include "json_transform.hpp"
+#include "duckdb/common/multi_file/multi_file_list.hpp"
 
 namespace duckdb {
 
@@ -142,6 +143,7 @@ void JSONScan::AddReadJSONParameters(TableFunction &table_function) {
 	table_function.named_parameters["timestampformat"] = LogicalType::VARCHAR;
 	table_function.named_parameters["timestamp_format"] = LogicalType::VARCHAR;
 	table_function.named_parameters["records"] = LogicalType::VARCHAR;
+	table_function.named_parameters["array"] = LogicalType::BOOLEAN;
 	table_function.named_parameters["maximum_sample_files"] = LogicalType::BIGINT;
 }
 
@@ -150,6 +152,375 @@ void JSONScan::AddAutoDetectParameters(TableFunction &table_function) {
 	table_function.named_parameters["field_appearance_threshold"] = LogicalType::DOUBLE;
 	table_function.named_parameters["convert_strings_to_integers"] = LogicalType::BOOLEAN;
 	table_function.named_parameters["map_inference_threshold"] = LogicalType::BIGINT;
+}
+
+bool JSONScan::ParseOption(ClientContext &context, const Identifier &key, const Value &value,
+                           JSONReaderOptions &options) {
+	if (value.IsNull()) {
+		throw BinderException("Cannot use NULL as argument to key %s", key);
+	}
+	if (key == "ignore_errors") {
+		options.ignore_errors = BooleanValue::Get(value);
+		return true;
+	}
+	if (key == "maximum_object_size") {
+		options.maximum_object_size = MaxValue<idx_t>(UIntegerValue::Get(value), options.maximum_object_size);
+		return true;
+	}
+	if (key == "format") {
+		auto arg = StringUtil::Lower(StringValue::Get(value));
+		static const auto FORMAT_OPTIONS =
+		    case_insensitive_map_t<JSONFormat> {{"auto", JSONFormat::AUTO_DETECT},
+		                                        {"unstructured", JSONFormat::UNSTRUCTURED},
+		                                        {"newline_delimited", JSONFormat::NEWLINE_DELIMITED},
+		                                        {"nd", JSONFormat::NEWLINE_DELIMITED},
+		                                        {"array", JSONFormat::ARRAY}};
+		auto lookup = FORMAT_OPTIONS.find(arg);
+		if (lookup == FORMAT_OPTIONS.end()) {
+			vector<string> valid_options;
+			for (auto &pair : FORMAT_OPTIONS) {
+				valid_options.push_back(StringUtil::Format("'%s'", pair.first));
+			}
+			throw BinderException("format must be one of [%s], not '%s'", StringUtil::Join(valid_options, ", "), arg);
+		}
+		options.format = lookup->second;
+		options.format_specified = true;
+		return true;
+	}
+	if (key == "compression") {
+		options.compression = FileCompressionType(StringValue::Get(value));
+		return true;
+	}
+	if (key == "array") {
+		// "array" is a shorthand for the format - a JSON array of values, or newline-delimited values
+		options.format = BooleanValue::Get(value) ? JSONFormat::ARRAY : JSONFormat::NEWLINE_DELIMITED;
+		options.format_specified = true;
+		return true;
+	}
+	if (key == "columns") {
+		auto &child_type = value.type();
+		if (child_type.id() != LogicalTypeId::STRUCT) {
+			throw BinderException("read_json \"columns\" parameter requires a struct as input.");
+		}
+		auto &struct_children = StructValue::GetChildren(value);
+		D_ASSERT(StructType::GetChildCount(child_type) == struct_children.size());
+		for (idx_t i = 0; i < struct_children.size(); i++) {
+			auto &name = StructType::GetChildName(child_type, i);
+			auto &val = struct_children[i];
+			if (val.IsNull()) {
+				throw BinderException("read_json \"columns\" parameter type specification cannot be NULL.");
+			}
+			options.name_list.emplace_back(name);
+			if (val.type().id() != LogicalTypeId::VARCHAR) {
+				throw BinderException("read_json \"columns\" parameter type specification must be VARCHAR.");
+			}
+			options.sql_type_list.emplace_back(TransformStringToLogicalType(StringValue::Get(val), context));
+		}
+		D_ASSERT(options.name_list.size() == options.sql_type_list.size());
+		if (options.name_list.empty()) {
+			throw BinderException("read_json \"columns\" parameter needs at least one column.");
+		}
+		return true;
+	}
+	if (key == "auto_detect") {
+		options.auto_detect = BooleanValue::Get(value);
+		options.auto_detect_specified = true;
+		return true;
+	}
+	if (key == "geojson") {
+		options.geojson = BooleanValue::Get(value);
+		return true;
+	}
+	if (key == "sample_size") {
+		auto arg = BigIntValue::Get(value);
+		if (arg == -1) {
+			options.sample_size = NumericLimits<idx_t>::Maximum();
+		} else if (arg > 0) {
+			options.sample_size = arg;
+		} else {
+			throw BinderException("read_json \"sample_size\" parameter must be positive, or -1 to sample all input "
+			                      "files entirely, up to \"maximum_sample_files\" files.");
+		}
+		return true;
+	}
+	if (key == "maximum_depth") {
+		auto arg = BigIntValue::Get(value);
+		if (arg == -1) {
+			options.max_depth = NumericLimits<idx_t>::Maximum();
+		} else {
+			options.max_depth = arg;
+		}
+		return true;
+	}
+	if (key == "field_appearance_threshold") {
+		auto arg = DoubleValue::Get(value);
+		if (arg < 0 || arg > 1) {
+			throw BinderException("read_json_auto \"field_appearance_threshold\" parameter must be between 0 and 1");
+		}
+		options.field_appearance_threshold = arg;
+		return true;
+	}
+	if (key == "map_inference_threshold") {
+		auto arg = BigIntValue::Get(value);
+		if (arg == -1) {
+			options.map_inference_threshold = NumericLimits<idx_t>::Maximum();
+		} else if (arg >= 0) {
+			options.map_inference_threshold = arg;
+		} else {
+			throw BinderException("read_json_auto \"map_inference_threshold\" parameter must be 0 or positive, "
+			                      "or -1 to disable map inference for consistent objects.");
+		}
+		return true;
+	}
+	if (key == "dateformat" || key == "date_format") {
+		auto format_string = StringValue::Get(value);
+		if (StringUtil::Lower(format_string) == "iso") {
+			format_string = "%Y-%m-%d";
+		}
+		options.date_format = format_string;
+
+		StrpTimeFormat format;
+		auto error = StrTimeFormat::ParseFormatSpecifier(format_string, format);
+		if (!error.empty()) {
+			throw BinderException("read_json could not parse \"dateformat\": '%s'.", error.c_str());
+		}
+		return true;
+	}
+	if (key == "timestampformat" || key == "timestamp_format") {
+		auto format_string = StringValue::Get(value);
+		if (StringUtil::Lower(format_string) == "iso") {
+			format_string = "%Y-%m-%dT%H:%M:%S.%fZ";
+		}
+		options.timestamp_format = format_string;
+
+		StrpTimeFormat format;
+		auto error = StrTimeFormat::ParseFormatSpecifier(format_string, format);
+		if (!error.empty()) {
+			throw BinderException("read_json could not parse \"timestampformat\": '%s'.", error.c_str());
+		}
+		return true;
+	}
+	if (key == "records") {
+		auto arg = StringValue::Get(value);
+		if (arg == "auto") {
+			options.record_type = JSONRecordType::AUTO_DETECT;
+		} else if (arg == "true") {
+			options.record_type = JSONRecordType::RECORDS;
+		} else if (arg == "false") {
+			options.record_type = JSONRecordType::VALUES;
+		} else {
+			throw BinderException("read_json requires \"records\" to be one of ['auto', 'true', 'false'].");
+		}
+		return true;
+	}
+	if (key == "maximum_sample_files") {
+		auto arg = BigIntValue::Get(value);
+		if (arg == -1) {
+			options.maximum_sample_files = NumericLimits<idx_t>::Maximum();
+		} else if (arg > 0) {
+			options.maximum_sample_files = arg;
+		} else {
+			throw BinderException("read_json \"maximum_sample_files\" parameter must be positive, or -1 to remove "
+			                      "the limit on the number of files used to sample \"sample_size\" rows.");
+		}
+		return true;
+	}
+	if (key == "convert_strings_to_integers") {
+		options.convert_strings_to_integers = BooleanValue::Get(value);
+		return true;
+	}
+	return false;
+}
+
+void JSONScan::BindSchema(ClientContext &context, JSONScanData &json_data, MultiFileList &files,
+                          vector<shared_ptr<JSONReader>> &sampled_readers, vector<LogicalType> &return_types,
+                          vector<Identifier> &names) {
+	auto &options = json_data.options;
+	names = options.name_list;
+	return_types = options.sql_type_list;
+	if (options.record_type == JSONRecordType::AUTO_DETECT && return_types.size() > 1) {
+		// More than one specified column implies records
+		options.record_type = JSONRecordType::RECORDS;
+	}
+
+	// Specifying column names overrides auto-detect
+	if (!return_types.empty()) {
+		options.auto_detect = false;
+	}
+
+	if (!options.auto_detect) {
+		// Need to specify columns if RECORDS and not auto-detecting
+		if (return_types.empty()) {
+			throw BinderException("When auto_detect=false, read_json requires columns to be specified through the "
+			                      "\"columns\" parameter.");
+		}
+		// If we are reading VALUES, we can only have one column
+		if (options.record_type == JSONRecordType::VALUES && return_types.size() != 1) {
+			throw BinderException("read_json requires a single column to be specified through the \"columns\" "
+			                      "parameter when \"records\" is set to 'false'.");
+		}
+	}
+
+	// Reading GeoJSON is opt-in, but a .geojson / .geojsonl file name is opt-in enough
+	if (!options.geojson.has_value()) {
+		const auto first_file = files.GetFirstFile().path;
+		options.geojson =
+		    StringUtil::CIEndsWith(first_file, ".geojson") || StringUtil::CIEndsWith(first_file, ".geojsonl");
+	}
+	// COPY ... FROM hardcodes RECORDS because it takes its columns from the target table. For GeoJSON we still
+	// need to know whether those columns live in a Feature's "properties", so let the structure decide.
+	if (options.geojson.value() && !options.auto_detect && !options.name_list.empty() &&
+	    options.record_type == JSONRecordType::RECORDS) {
+		options.record_type = JSONRecordType::AUTO_DETECT;
+	}
+
+	json_data.InitializeFormats();
+
+	if (options.auto_detect || options.record_type == JSONRecordType::AUTO_DETECT) {
+		JSONScan::AutoDetect(context, json_data, files.GetAllFiles(), sampled_readers, return_types, names);
+		D_ASSERT(return_types.size() == names.size());
+	}
+	json_data.key_names = IdentifiersToStrings(names);
+}
+
+void JSONScan::FinalizeBind(JSONScanData &json_data, vector<Identifier> &names) {
+	auto &options = json_data.options;
+	auto &transform_options = json_data.transform_options;
+	transform_options.strict_cast = !options.ignore_errors;
+	transform_options.error_duplicate_key = !options.ignore_errors;
+	transform_options.error_missing_key = false;
+	transform_options.error_unknown_key = options.auto_detect && !options.ignore_errors;
+	transform_options.date_format_map = json_data.date_format_map.get();
+	transform_options.delay_error = true;
+
+	if (!options.auto_detect) {
+		return;
+	}
+	// Note that we can't change json_data.key_names, because the JSON reader gets columns by exact name, not position
+	DeduplicateColumnNames(names);
+}
+
+void JSONScan::DeduplicateColumnNames(vector<Identifier> &names) {
+	identifier_map_t<idx_t> name_collision_count;
+	for (auto &col_name : names) {
+		// Taken from CSV header_detection.cpp
+		while (name_collision_count.find(col_name) != name_collision_count.end()) {
+			name_collision_count[col_name] += 1;
+			col_name = Identifier(col_name + "_" + to_string(name_collision_count[col_name]));
+		}
+		name_collision_count[col_name] = 0;
+	}
+}
+
+//! GeoJSON Features are read as two groups of columns: those taken from the Feature itself, and those taken from
+//! its "properties" object
+static bool TransformGeoJSONFeatures(yyjson_val *values[], yyjson_alc *alc, const idx_t count,
+                                     JSONScanGlobalState &gstate, const vector<Vector *> &result_vectors,
+                                     JSONTransformOptions &transform_options) {
+	D_ASSERT(gstate.feature_columns.size() == gstate.names.size());
+
+	// Group the columns by where they read from, using the JSON key rather than the (deduplicated) column name.
+	// A Feature member and a property of the same name are separate columns, so this cannot be done by name.
+	vector<string> feature_names, property_names;
+	vector<Vector *> feature_vectors, property_vectors;
+	vector<ColumnIndex> feature_indices, property_indices;
+	for (idx_t i = 0; i < gstate.feature_columns.size(); i++) {
+		const auto &feature_column = gstate.feature_columns[i];
+		const auto from_properties = feature_column.from_properties;
+		(from_properties ? property_names : feature_names).push_back(feature_column.key);
+		(from_properties ? property_vectors : feature_vectors).push_back(result_vectors[i]);
+		if (!gstate.column_indices.empty()) {
+			(from_properties ? property_indices : feature_indices).push_back(gstate.column_indices[i]);
+		}
+	}
+
+	// "type", "properties" and "bbox" are consumed by the unnesting, so unknown keys are expected in both groups
+	bool success = true;
+	if (!feature_names.empty()) {
+		success = JSONTransform::TransformObject(values, alc, count, feature_names, feature_vectors, transform_options,
+		                                         feature_indices, false);
+	}
+	if (success && !property_names.empty()) {
+		vector<yyjson_val *> property_values(count);
+		for (idx_t i = 0; i < count; i++) {
+			auto &val = values[i];
+			property_values[i] = val && unsafe_yyjson_is_obj(val) ? yyjson_obj_get(val, "properties") : nullptr;
+		}
+		success = JSONTransform::TransformObject(property_values.data(), alc, count, property_names, property_vectors,
+		                                         transform_options, property_indices, false);
+	}
+	return success;
+}
+
+void ReadJSONFunction(ClientContext &context, JSONReader &json_reader, JSONScanGlobalState &gstate,
+                      JSONScanLocalState &lstate, DataChunk &output) {
+	auto &scan_state = lstate.GetScanState();
+	D_ASSERT(RefersToSameObject(json_reader, *scan_state.current_reader));
+
+	const auto count = lstate.Read();
+	yyjson_val **values = scan_state.values;
+
+	if (!gstate.names.empty()) {
+		vector<Vector *> result_vectors;
+		result_vectors.reserve(gstate.names.size());
+		for (idx_t i = 0; i < gstate.names.size(); i++) {
+			result_vectors.emplace_back(&output.data[i]);
+		}
+
+		D_ASSERT(gstate.json_data.options.record_type != JSONRecordType::AUTO_DETECT);
+		bool success;
+		if (gstate.json_data.options.record_type == JSONRecordType::RECORDS) {
+			success = JSONTransform::TransformObject(values, scan_state.allocator.GetYYAlc(), count, gstate.names,
+			                                         result_vectors, lstate.transform_options, gstate.column_indices,
+			                                         lstate.transform_options.error_unknown_key);
+		} else if (gstate.json_data.options.record_type == JSONRecordType::FEATURES) {
+			success = TransformGeoJSONFeatures(values, scan_state.allocator.GetYYAlc(), count, gstate, result_vectors,
+			                                   lstate.transform_options);
+		} else {
+			D_ASSERT(gstate.json_data.options.record_type == JSONRecordType::VALUES);
+			success = JSONTransform::Transform(values, scan_state.allocator.GetYYAlc(), *result_vectors[0], count,
+			                                   lstate.transform_options, gstate.column_indices[0]);
+		}
+
+		if (!success) {
+			string hint =
+			    gstate.json_data.options.auto_detect
+			        ? "\nTry increasing 'sample_size', reducing 'maximum_depth', specifying 'columns', 'format' or "
+			          "'records' manually, setting 'ignore_errors' to true, or setting 'union_by_name' to true when "
+			          "reading multiple files with a different structure."
+			        : "\nTry setting 'auto_detect' to true, specifying 'format' or 'records' manually, or setting "
+			          "'ignore_errors' to true.";
+			lstate.AddTransformError(lstate.transform_options.object_index,
+			                         lstate.transform_options.error_message + hint);
+			return;
+		}
+	}
+	output.SetChildCardinality(count);
+}
+
+void ReadJSONObjectsFunction(ClientContext &context, JSONReader &json_reader, JSONScanGlobalState &gstate,
+                             JSONScanLocalState &lstate, DataChunk &output) {
+	// Fetch next lines
+	auto &scan_state = lstate.GetScanState();
+	D_ASSERT(RefersToSameObject(json_reader, *scan_state.current_reader));
+
+	const auto count = lstate.Read();
+	const auto units = scan_state.units;
+	const auto objects = scan_state.values;
+
+	if (!gstate.names.empty()) {
+		// Create the strings without copying them
+		auto strings = FlatVector::Writer<string_t>(output.data[0], count);
+		for (idx_t i = 0; i < count; i++) {
+			if (objects[i]) {
+				strings.WriteStringRef(string_t(units[i].pointer, units[i].size));
+			} else {
+				strings.WriteNull();
+			}
+		}
+	}
+
+	output.SetChildCardinality(count);
 }
 
 } // namespace duckdb
